Backup and copy the images directory over a tmpfs directory
The idea here is if a user chooses to use a tmpfs over a directory, then we
want the content of the images directory to be preserved.  The first use
case I would like to use for this would be /run.  But if the --tmpfs patch or
similar got merged, I think you would need something like this patch.

Docker-DCO-1.1-Signed-off-by: Dan Walsh <dwalsh@redhat.com> (github: rhatdan)