set custom namespaces for init process
An init process can join other namespaces (pidns, ipc etc.). This leverages
C code defined in nsenter package to spawn a process with correct namespaces
and clone if necessary.  When a shared-pidns container exits,
the original container will keep running.

This removes joinExistingNamespaces in Go layer because all namespaces will be
setns in C layer.

This also changes setns process to requires a list of namespaces to be joined
rather than requires only the pid of init process and deduct the namespace from
it.

Signed-off-by: Daniel, Dao Quang Minh <dqminh89@gmail.com>