cgroups: fs: fix issues with .Apply()
If a configuration option is set for a cgroup that isn't mounted, an
error will be emitted. This is an issue, because newer cgroups aren't
present in older kernels, and we can't break backcompat where a new
config can't run on an older kernel.

The issue arises because .Set() does no checks for whether the given
path actually points to a mountpoint, so fix the code by just bailing if
the cgroup mountpoint doesn't exist.

Fixes: fc3981ea5c10
Fixes: 606d9064b0a6
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>